Commit - 202210101118
Updated README.md

Task 9 - Added the US and Spain vector images, to be used for the country image, to the project. Task not yet complete.

Task 3 - Added button colors for both the english and spanish languages in the colors.xml file(s).

(Build.gradle) Added view binding and build features to allow for binding in MainActivity.java. Had a little trouble with adding this as I had to add several things that were not present in previous assignments.

(MainActivity.java) Added the binding to the MainActivity class.