Commit - 202210101234
Updated README.md

Task 4 - Added the country descriptions for Spain and USA in their respective strings.xml files. Added the loading of the country description text to the changeCountryDescription() method in the MainActivity.java file.

Todo 5 - Initialized in the onCreate() method to its own function, changeCountryDescription().

Added a button listener for the onLessonLearntClicked() method.