Commit - 202210101452
Updated README.md

Remove extraneous comments

Task 6 & 7 - Added the button_default.xml to define to define the state the button will be in when the user has not pressed the button and created button_pressed.xml for when they have. Also, added the button_color.xml as a selector to differentiate which of the xml files is to be used when the user presses the button on the screen. Changed the button AppTheme to use a drawable resource instead of a color. This will allow themes to be changed when the language changes in the app.