Commit - 202210101658
Updated README.md

Added the country vector images that are tied to the language of the system. Now, the appropriate country image will populate the country_image ImageView dependent upon the users chosen language. Also, removed the older versions of the vector images as the implementation changed.

(MainActivity.java) - Added coordinator layout and implemented a snackbar when the user clicks the "lesson learnt" button.